requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyMySQL==1.1.1
cryptography>=42.0.0
openai>=1.35.0
anthropic>=0.31.0
google-generativeai>=0.7.2
Werkzeug>=3.0.3
itsdangerous>=2.2.0
Jinja2>=3.1.4
requests>=2.32.3
pydantic>=2.8.2
xhtml2pdf>=0.2.15
